[MIG] Add improves from PR 265: Migration to 17.0
