+++ loyalty_partner_applicability/readme/CONTRIBUTORS.md
@@ -1,2 +1,4 @@
 - [Tecnativa](https://www.tecnativa.com):
   - Pilar Vargas
+- [ACSONE SA/NV](https://acsone.eu):
+  - Laurent Mignon <laurent.mignon@acsone.eu>
