+++ loyalty_partner_applicability/readme/DESCRIPTION.md
@@ -1,3 +1,9 @@
-This module extends the Loyalty functionality by allowing the definition
-of a customer filter for promotion rules. This module is a base to be
-extended.
+This module extends the Loyalty functionality by allowing the definition of a customer
+filter for promotion rules. This module is a base to be extended.
+
+It allows to restrict the applicability of a promotion and its rules to a subset of
+customers. This is useful when you want to target a specific group of customers with a
+promotion. The restriction can be defined at the program level.
+
+When a restriction is defined at the program level, the promotion will only be
+available to customers that match the filter.
