+++ loyalty_partner_applicability/readme/USAGE.md
@@ -5,5 +5,5 @@ Restriction on partners can be expressed in the following ways:
 
 If both are provided, the condition will be a logical OR.
 
-When you define a domain the datetime object is available as a variable
-datetime to allow for dynamic filtering based on the current datetime.
+When you define a domain the datetime object is available as a variable `datetime` to
+allow for dynamic filtering based on the current datetime.
