+++ loyalty_partner_applicability/tests/test_loyalty_partner_applicability.py
@@ -49,22 +49,3 @@ class TestLoyaltyPartnerApplicability(TestLoyaltyPartnerApplicabilityCase):
         self._assertCheckValidPartner(program, self.partner1, True)
         self._assertCheckValidPartner(program, self.partner2, True)
         self._assertCheckValidPartner(program, self.partner3, False)
-
-    def test_restriction_on_partner_at_rule_level(self):
-        program = self.program_and_rule_restricted
-        # At program level the partner 1 and 2 are valid
-        self.assertFalse(program._is_coupon_sharing_allowed())
-        self._assertCheckValidPartner(program, self.partner1, True)
-        self._assertCheckValidPartner(program, self.partner2, True)
-        self._assertCheckValidPartner(program, self.partner3, False)
-        # At rule level we have 1 rule for partner 1 and 2 and 1 rule for partner 2
-        rule = program.rule_ids.filtered(lambda r: r._is_partner_valid(self.partner1))
-        self.assertEqual(len(rule), 1)
-        rule = program.rule_ids.filtered(lambda r: r._is_partner_valid(self.partner2))
-        self.assertEqual(len(rule), 1)
-        # even if a rule is allowed for partner 3, it's not valid since
-        # the program is not valid for partner 3
-        rule = program.rule_ids.filtered(lambda r: self.partner3 in r.partner_ids)
-        self.assertEqual(len(rule), 1)
-        rule = program.rule_ids.filtered(lambda r: r._is_partner_valid(self.partner3))
-        self.assertEqual(len(rule), 0)
